requirements.txt
python-telegram-bot==13.15
openai>=1.65.0
requests==2.31.0
python-dotenv==1.0.0
notion-client==2.0.0
